% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quoll.R
\name{community}
\alias{community}
\title{Define a signed community matrix}
\usage{
community(defn)
}
\arguments{
\item{defn}{a sequence of statements describing the network model}
}
\value{
a signed adjacency matrix
}
\description{
Symbolically specify a directed graph representing a network
model.
}
\details{
Generate a signed adjacency matrix from a sequence of symbolic
directives that describe the network model.  The directives are:
\itemize{
\item \code{vertices(v1,v2,...)} defines the vertices of the graph.
\item \code{loop(v1,v2,...)} defines self loops for the specified
vertices.
\item \code{a \%->\% b} defines a unidirectional edge where \code{a} has a
positive impact on \code{b}
\item \code{a \%-o\% b} defines a unidirectional edge where \code{a} has a
negative impact on \code{b}
\item \code{a \%<-\% b} defines a unidirectional edge where \code{b} has a
positive impact on \code{a}
\item \code{a \%o-\% b} defines a unidirectional edge where \code{b} has a
negative impact on \code{a}
\item \code{a \%o->\% b} is equivalent to \code{a \%0-\% b} and \code{a \%->\% b}
\item \code{a \%<-o\% b} is equivalent to \code{a \%<-\% b} and \code{a \%-o\% b}
\item \code{a \%<->\% b} is equivalent to \code{a \%<-\% b} and \code{a \%->\% b}
\item \code{a \%<->\% b} is equivalent to \code{a \%o-\% b} and \code{a \%-o\% b}
}
}
\examples{
# Define a signed community matrix
S <- community({
  vertices(A,B,C,D,E)
  loop(A,B,C,D,E)
  A \%->\% B \%->\% D
  A \%->\% C \%->\% D
  B \%o->\% E
})
S
}
\references{
Dambacher, J. M., Li, H. W., & Rossignol,
P. A. (2002).  Relevance of community structure in assessing
indeterminacy of ecological predictions. \emph{Ecology}, 83(5),
1372-1385.
}
