% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quoll.R
\name{community_diagram}
\alias{community_diagram}
\title{Network diagram}
\usage{
community_diagram(
  S,
  rankdir = c("BT", "TB", "LR", "RL"),
  vertex.col = colorspec_vector(S),
  edge.col = colorspec_matrix(S)
)
}
\arguments{
\item{S}{a signed community matrix}

\item{rankdir}{the layout direction of the graph}

\item{vertex.col}{a colorspec vector of vertex colors}

\item{edge.col}{a colorspec matrix of edge colors}
}
\value{
a dot description of the graph
}
\description{
Create a diagram of the directed graph representation of a
network model.
}
\details{
Creates a dot description of the directed graph representation of
a network model, to be rendered with \code{\link[DiagrammeR:grViz]{DiagrammeR::grViz()}}.
}
