% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quoll.R
\name{connected_components}
\alias{connected_components}
\title{Connected components}
\usage{
connected_components(A)
}
\arguments{
\item{A}{an adjacency matrix of a graph}
}
\value{
a list of vectors of the vertices in each connected component
}
\description{
Find the connected components of a graph
}
\details{
Returns a list of the connected components of a graph, represented
as vectors of vertices.
}
\examples{
# Adjacency matrix of a directed graph
A <- matrix(c(1,1,0,0,1,1,0,0,0,0,1,1,0,0,1,1),4,4)
# Find all connected components
connected_components(A)
}
