% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quoll.R
\name{cpoly}
\alias{cpoly}
\title{Characteristic polynomial}
\usage{
cpoly(S, factor = TRUE)
}
\arguments{
\item{S}{a signed adjacency matrix}

\item{factor}{should the characteristic polynomial be factored}
}
\value{
a \code{cpoly} object
}
\description{
Characteristic polynomial of a matrix
}
\details{
Given \code{S=sign(X)}, \code{cpoly} generates a symbolic representation of
the characteristic polynomial of the matrix \code{X}. The functions
\code{\link[=cpoly_latex]{cpoly_latex()}} and \code{\link[=cpoly_R]{cpoly_R()}} render this symbolic
representation as LaTeX and R code.

If \code{factor=TRUE}, the characteristic polynomial is factored into
separate components, one for each strongly connected component of
the corresponding graph.
}
\examples{
# Tridiagonal sign matrix
n <- 3
S <- diag(-1,n,n)+rbind(0,diag(1,n-1,n))+cbind(0,diag(1,n,n-1))
# Generate R function for coeffs of the characteristic polynomial
cat(cpoly_R(cpoly(S)))
# Generate latex for the characteristic polynomial
cat(cpoly_latex(cpoly(S)))
}
\seealso{
\code{\link[=cpoly_latex]{cpoly_latex()}}, \code{\link[=cpoly_R]{cpoly_R()}}
}
