% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quoll.R
\name{cpoly_components}
\alias{cpoly_components}
\title{Render components of a characteristic polynomial}
\usage{
cpoly_components(cpoly, fmt)
}
\arguments{
\item{cpoly}{a \code{cpoly} object}

\item{fmt}{a list of format functions and strings}
}
\value{
a list of character vectors that define
\itemize{
\item \code{cyc} assignments to cycle weight variables
\item \code{cft} lists of assignments to polynomial coefficients, one for each component
}
}
\description{
Render components of a \code{cpoly} object
}
\details{
This is the workhorse function that generates expressions to
evaluate the coeffcients of the characteristic polynomial.

The \code{fmt} argument is a list of functions and strings that
construct syntactic elements of the generated code:
\itemize{
\item \code{asub(i,j)} function to generate the i,j element of the weight matrix
\item \code{csub(k)} function to generate the k-th cycle weight variable
\item \code{bsub(k)} function to generate the k-th polynomial coefficient variable
\item \code{cassign(l,r)} vectorized function to generate cycle weight assignments (l=r)
\item \code{bassign(l,r)} vectorized function to generate polynomial coefficient assignments (l=r)
\item \code{mul} string for the multiplication operator
}
}
\seealso{
\code{\link[=cpoly]{cpoly()}}
}
