% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quoll.R
\name{cpoly_latex}
\alias{cpoly_latex}
\title{Render a \code{cpoly} object as LaTeX}
\usage{
cpoly_latex(cpoly, use.names = FALSE)
}
\arguments{
\item{cpoly}{a \code{cpoly} object}

\item{use.names}{should variable names be used as subscripts}
}
\value{
a character string of LaTeX code
}
\description{
Render a \code{cpoly} object as a LaTeX expression for the
characteristic polynomial.
}
\details{
This function translates a \code{cpoly} object generated by \code{\link[=cpoly]{cpoly()}}
to LaTeX expressions for the characteristic polynomial.  In the
generated LaTeX,
\itemize{
\item the \code{a} are (unsigned) edge weights,
\item the \code{c} are cycle weights,
\item the \code{b} are coefficients of the characteristic polynomial.
}
}
\seealso{
\code{\link[=cpoly]{cpoly()}}
}
