% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quoll.R
\name{cpoly_R}
\alias{cpoly_R}
\title{Render cpoly object as R code}
\usage{
cpoly_R(cpoly)
}
\arguments{
\item{cpoly}{a \code{cpoly} object}
}
\value{
a character string of R code
}
\description{
Render a \code{cpoly} object as R code to evaluate the
coefficients of the characteristic polynomial.
}
\details{
This function translates a \code{cpoly} object generated by \code{\link[=cpoly]{cpoly()}}
to R code to evaluate the coefficients of the characteristic
polynomial.  In the generated R code,
\itemize{
\item the argument \code{A} is a matrix of (unsigned) edge weights,
\item the \code{C} variables are cycle weights,
\item the \code{B} variables are coefficients of the characteristic polynomial.
}
}
\seealso{
\code{\link[=cpoly]{cpoly()}}
}
