% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quoll.R
\name{determinant_latex}
\alias{determinant_latex}
\title{Render determinant as LaTeX}
\usage{
determinant_latex(determinant, use.names = FALSE)
}
\arguments{
\item{determinant}{a determinant object}

\item{use.names}{should variable names be used as subscripts}
}
\value{
a character string of LaTeX code.
}
\description{
Render a \code{partial} object as LaTeX expressions for the determinant.
}
\details{
This function translates a \code{partial} object generated by
\code{\link[=partial]{partial()}} to LaTeX code for the determinant.  In the generated LaTeX,
\itemize{
\item the \code{a} are (unsigned) edge weights,
\item the \code{c} are cycle weights,
\item the \code{d} are determinants of strongly connected components
}

The determinant of the matrix (d_0) is the product of the determinants of
the strongly connected components.
}
\seealso{
\code{\link[=partial]{partial()}}
}
