% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quoll.R
\name{determinant_R}
\alias{determinant_R}
\title{Render determinant as R code}
\usage{
determinant_R(determinant)
}
\arguments{
\item{determinant}{a determinant object}
}
\value{
a character string of R code.
}
\description{
Render a \code{partial} object as R code to calculate the determinant.
}
\details{
This function translates a \code{partial} object generated by
\code{\link[=partial]{partial()}} to R code to evaluate the determinant.  In the
generated code,
\itemize{
\item the argument \code{A} is a matrix of (unsigned) edge weights,
\item the \code{C} variables are cycle weights,
\item the \code{D} variables are determinants of strongly connected components.
}

The determinant of the matrix is the product of the determinants
of the strongly connected components.
}
\seealso{
\code{\link[=partial]{partial()}}
}
