% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quoll.R
\name{digraph}
\alias{digraph}
\title{Define an adjacency matrix}
\usage{
digraph(defn)
}
\arguments{
\item{defn}{a sequence of statements describing the digraph}
}
\value{
an adjacency matrix
}
\description{
Symbolically specify a directed graph.
}
\details{
Generate an adjacency matrix from a sequence of symbolic
directives that describe a directed graph.  The directives are:
\itemize{
\item \code{vertices(v1,v2,...)} defines the vertices of the graph.
\item \code{a \%->\% b} defines a unidirectional edge from \code{a} to \code{b}
\item \code{a \%<-\% b} defines a unidirectional edge from \code{b} to \code{a}
\item \code{a \%<->\% b} is equivalent to \code{a \%<-\% b} and \code{a \%->\% b}
}
}
\examples{
# Define a directed graph
A <- digraph({
  vertices(A,B,C,D,E)
  A \%->\% B \%->\% D
  A \%->\% C \%->\% D
  B \%<-\% E
})
A
}
