% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quoll.R
\name{directed_cycles}
\alias{directed_cycles}
\title{Cycles of a directed graph}
\usage{
directed_cycles(A, loops = TRUE)
}
\arguments{
\item{A}{adjacency matrix of a directed graph}

\item{loops}{should self loops be included}
}
\value{
a list of cycles
}
\description{
Find all cycles in a directed graph, optionally including self
loops.
}
\details{
Returns a list of the cycles of a directed graph, optionally
including self loops.  Cycles are represented as vectors of
vertices, lowest numbered vertex first.
}
\examples{
# Adjacency matrix of a directed graph
A <- matrix(c(1,1,0,0,1,1,1,0,0,0,1,1,1,0,1,1),4,4)
# Find all cycles in the graph, excluding loops
directed_cycles(A,loops=FALSE)
}
