% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quoll.R
\name{eval_poly}
\alias{eval_poly}
\title{Evaluate a polynomial}
\usage{
eval_poly(cs, x)
}
\arguments{
\item{cs}{a vector of polynomial coefficients}

\item{x}{a vector or matrix of values}
}
\value{
a vector or matrix of polynomial values
}
\description{
Evaluate a polynomial
}
\details{
This function evaluates a polynomial with coefficients \code{cs} at
values \code{x}.  If \code{x} is a matrix, then \code{x} must be square and the
matrix polynomial is evaluated.
}
