% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quoll.R
\name{find_paths}
\alias{find_paths}
\title{Paths in a directed graph}
\usage{
find_paths(A, from, to)
}
\arguments{
\item{A}{adjacency matrix of a directed graph}

\item{from}{initial vertex of path}

\item{to}{final vertex of path}
}
\value{
a list of paths, represented as a vector of vertices
}
\description{
Find all paths between two vertices, ignoring self loops
}
\examples{
# Adjacency matrix of a directed graph
A <- matrix(c(1,1,0,0,1,1,1,0,0,0,1,1,1,0,1,1),4,4)
# Find all paths from vertex 1 to vertex 4
find_paths(A,1,4)
}
