% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quoll.R
\name{highlight_ncycle}
\alias{highlight_ncycle}
\title{Highlight negative cycles in LaTeX}
\usage{
highlight_ncycle(latex, col = "red")
}
\arguments{
\item{latex}{a character string of LaTeX code}

\item{col}{the highlight colour}
}
\value{
a character string of LaTeX code
}
\description{
Highlight negative cycle weights in LaTeX
}
\details{
The \code{highlight_negative} function highlights negative cycle
weights in LaTeX code generated by \code{\link[=partial_latex]{partial_latex()}}.
}
\seealso{
\code{\link[=partial_latex]{partial_latex()}}
}
