% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quoll.R
\name{intersection_graph}
\alias{intersection_graph}
\alias{intersection_graph_int}
\title{Intersection graph}
\usage{
intersection_graph(sets)

intersection_graph_int(sets, maxel)
}
\arguments{
\item{sets}{a list of vectors, each representing a set}

\item{maxel}{the maximum element of any set}
}
\value{
the adjacency matrix of the intersection graph
}
\description{
Create the intersection graph of a list of sets
}
\details{
Creates an adjacency matrix representing the intersection graph of
a collection of sets.  In the intersection graph, vertices
represent sets and edges connect sets that have a non-empty
intersection.

Two implementations are provided:
\itemize{
\item intersection_graph() - works with sets of any type
\item intersection_graph_int() - optimized for integer sets
}
}
\examples{
# List of sets
sets <- list(c(1,2,3), c(2,3,4), c(3,4,5))
# Create the intersection graph
intersection_graph(sets)
}
