% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quoll.R
\name{maximal_IS}
\alias{maximal_IS}
\title{Maximal independent sets}
\usage{
maximal_IS(A, exclude = c(), include = c(), subgraph = 1:nrow(A))
}
\arguments{
\item{A}{adjacency matrix of a graph}

\item{exclude}{sets that include these vertices are excluded.}

\item{include}{only sets that include this vertex are included.}

\item{subgraph}{the subset of vertices to search}
}
\value{
a list of maximal independent sets
}
\description{
Find the maximal independent sets of a graph
}
\details{
Returns a list of the maximal independent sets of a subgraph of a
graph, represented as vectors of vertices. The indendependend sets
can be filtered to exclude set containing certain vertices, and to
include only sets containing a (single) specified vertex.
}
