% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quoll.R
\name{partial}
\alias{partial}
\alias{partialA}
\title{Partial matrix inverse}
\usage{
partial(S, i = integer(0), j = integer(0))

partialA(S, i = integer(0), j = integer(0))
}
\arguments{
\item{S}{a signed adjacency matrix}

\item{i}{vector of row indices}

\item{j}{vector of column indices}
}
\value{
a partial object for the partial inverse of a matrix with the
same sign structure as the signed adjacency matrix S.
}
\description{
Generate elements of the inverse of a matrix
}
\details{
Given \code{S=sign(X)}, \code{partial} generates a symbolic representation
for a specified subset of elements of the inverse or adjoint of \code{X},
expressed in terms of the elements of the weight matrix \code{A=abs(X)}.
The functions \code{\link[=partial_latex]{partial_latex()}}, \code{\link[=partial_R]{partial_R()}} and \code{\link[=partial_cpp]{partial_cpp()}}
render the symbolic representation returned by \code{partial} as LaTeX,
R or C++ code.

\code{partialA} is an alternative implementation of \code{partial} that is
often faster when many elements of the inverse or adjoint are
required.  \code{partial} and \code{partialA} index cycles differently, and
so can produce different but equivalent expressions.
}
\examples{
# Tridiagonal sign matrix
n <- 3
S <- diag(-1,n,n)+rbind(0,diag(1,n-1,n))+cbind(0,diag(1,n,n-1))
# Generate R function for all elements of the inverse
cat(partial_R(partial(S,row(S),col(S))))
# Generate latex for all elements of the inverse
cat(partial_latex(partial(S,row(S),col(S))))
}
\seealso{
\code{\link[=partial_latex]{partial_latex()}}, \code{\link[=partial_R]{partial_R()}}, \code{\link[=partial_cpp]{partial_cpp()}}
}
