% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quoll.R
\name{partial_latex}
\alias{partial_latex}
\title{Render partial object as LaTeX}
\usage{
partial_latex(
  partial,
  type = c("inv", "adj"),
  determinant = TRUE,
  use.names = FALSE
)
}
\arguments{
\item{partial}{a partial inverse object}

\item{type}{whether the inverse or adjoint is generated}

\item{determinant}{should uneccessary factors be retained}

\item{use.names}{should variable names be used as subscripts}
}
\value{
a character string of LaTeX code.
}
\description{
Render a \code{partial} object as LaTeX expressions for the partial
inverse or adjoint matrix.
}
\details{
This function translates a \code{partial} object generated by
\code{\link[=partial]{partial()}} to LaTeX code for the elements of the inverse or
adjoint matrix.  In the generated LaTeX,
\itemize{
\item the \code{a} are (unsigned) edge weights,
\item the \code{c} are cycle weights,
\item the \code{d} are determinants of strongly connected components
}

If \code{determinant=FALSE}, any \code{D} variables not required for the
subsequent calculation areomitted.  If \code{determinant=TRUE}, no
variables are omitted and the product of the \code{D} variables is the
determinant.
}
\seealso{
\code{\link[=partial]{partial()}}, \code{\link[=highlight_ncycle]{highlight_ncycle()}}
}
