% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quoll.R
\name{print.cpoly}
\alias{print.cpoly}
\title{Print characteristic polynomial}
\usage{
\method{print}{cpoly}(x, ...)
}
\arguments{
\item{x}{a characteristic polynomial}

\item{...}{ignored}
}
\description{
Print a characteristic polynomial object
}
