% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quoll.R
\name{print.partial}
\alias{print.partial}
\title{Print partial inverse}
\usage{
\method{print}{partial}(x, ...)
}
\arguments{
\item{x}{a partial inverse}

\item{...}{ignored}
}
\description{
Print a partial inverse object
}
