% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quoll.R
\name{reachable}
\alias{reachable}
\title{Reachable vertices}
\usage{
reachable(A, root)
}
\arguments{
\item{A}{an adjacency matrix of a directed graph}

\item{root}{the root vertex}
}
\value{
a vector of vertices reachable from the root
}
\description{
Find the vertices reachable from a root vertex in a directed graph
}
\examples{
# Adjacency matrix of a directed graph
A <- matrix(c(1,1,0,0,0,1,0,0,0,0,1,1,0,0,1,1),4,4)
# Find all vertices reachable from vertex 1
reachable(A,1)
}
