% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quoll.R
\name{render_path}
\alias{render_path}
\alias{render_cycle}
\alias{render_Cexpr}
\alias{render_product}
\alias{render_Dproduct}
\alias{render_wpath}
\alias{render_inverse_element}
\alias{render_adjoint_element}
\title{Render R utility functions}
\usage{
render_path(path, S, fmt)

render_cycle(cycle, S, fmt)

render_Cexpr(cexpr, fmt, sign = "+")

render_product(a, b, fmt)

render_Dproduct(dets, fmt)

render_wpath(term, S, fmt)

render_inverse_element(element, S, fmt)

render_adjoint_element(element, S, fmt)
}
\arguments{
\item{path}{a vector of vertices}

\item{S}{the signed adjacency matrix}

\item{fmt}{a list of format functions and strings}

\item{cycle}{a vector of vertices}

\item{cexpr}{a vector of cycle numbers}

\item{sign}{a sign character}

\item{a, b}{(unsigned) multiplicands}

\item{dets}{indices of component determinants}

\item{term}{a term of an element}

\item{element}{an element of a partial object}
}
\value{
a character string
}
\description{
Utility functions for rendering partial inverse objects
}
