% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quoll-package.R
\docType{package}
\name{quoll-package}
\alias{quoll}
\alias{quoll-package}
\title{quoll: Qualitiative Network Analysis}
\description{
Caclulate inverse and adjoint matrices by graph theoretic methods for qualitative network analyis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SWotherspoon/quoll}
  \item Report bugs at \url{https://github.com/SWotherspoon/quoll/issues}
}

}
\author{
\strong{Maintainer}: Simon Wotherspoon \email{Simon.Wotherspoon@aad.gov.au} (\href{https://orcid.org/0000-0002-6947-4445}{ORCID})

}
\keyword{internal}
